+++ pkgs/test/lib/scaffolding.dart
@@ -2,5 +2,4 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-// ignore: deprecated_member_use
-export 'package:test_core/scaffolding.dart';
+export 'package:test_core/src/scaffolding.dart';