+++ pkgs/test/lib/src/runner/browser/platform.dart
@@ -3,41 +3,26 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'dart:async';
-import 'dart:convert';
 import 'dart:io';
 
 import 'package:async/async.dart';
-import 'package:http_multi_server/http_multi_server.dart';
 import 'package:path/path.dart' as p;
-import 'package:pool/pool.dart';
-import 'package:shelf/shelf.dart' as shelf;
-import 'package:shelf/shelf_io.dart' as shelf_io;
-import 'package:shelf_packages_handler/shelf_packages_handler.dart';
-import 'package:shelf_static/shelf_static.dart';
-import 'package:shelf_web_socket/shelf_web_socket.dart';
-import 'package:test_api/backend.dart'
-    show Compiler, Runtime, StackTraceMapper, SuitePlatform;
+import 'package:test_api/backend.dart' show Compiler, Runtime, SuitePlatform;
 import 'package:test_core/src/runner/configuration.dart'; // ignore: implementation_imports
-import 'package:test_core/src/runner/dart2js_compiler_pool.dart'; // ignore: implementation_imports
 import 'package:test_core/src/runner/load_exception.dart'; // ignore: implementation_imports
-import 'package:test_core/src/runner/package_version.dart'; // ignore: implementation_imports
 import 'package:test_core/src/runner/platform.dart'; // ignore: implementation_imports
 import 'package:test_core/src/runner/plugin/customizable_platform.dart'; // ignore: implementation_imports
 import 'package:test_core/src/runner/runner_suite.dart'; // ignore: implementation_imports
 import 'package:test_core/src/runner/suite.dart'; // ignore: implementation_imports
-import 'package:test_core/src/util/errors.dart'; // ignore: implementation_imports
-import 'package:test_core/src/util/io.dart'; // ignore: implementation_imports
 import 'package:test_core/src/util/package_config.dart'; // ignore: implementation_imports
-import 'package:test_core/src/util/stack_trace_mapper.dart'; // ignore: implementation_imports
-import 'package:web_socket_channel/web_socket_channel.dart';
 import 'package:yaml/yaml.dart';
 
-import '../../util/math.dart';
-import '../../util/one_off_handler.dart';
-import '../../util/package_map.dart';
-import '../../util/path_handler.dart';
 import '../executable_settings.dart';
 import 'browser_manager.dart';
+import 'compilers/compiler_support.dart';
+import 'compilers/dart2js.dart';
+import 'compilers/dart2wasm.dart';
+import 'compilers/precompiled.dart';
 import 'default_settings.dart';
 
 class BrowserPlatform extends PlatformPlugin
@@ -47,60 +32,61 @@ class BrowserPlatform extends PlatformPlugin
   /// [root] is the root directory that the server should serve. It defaults to
   /// the working directory.
   static Future<BrowserPlatform> start({String? root}) async {
-    var server = shelf_io.IOServer(await HttpMultiServer.loopback(0));
     var packageConfig = await currentPackageConfig;
     return BrowserPlatform._(
-        server,
         Configuration.current,
         p.fromUri(packageConfig.resolve(
             Uri.parse('package:test/src/runner/browser/static/favicon.ico'))),
         p.fromUri(packageConfig.resolve(Uri.parse(
             'package:test/src/runner/browser/static/default.html.tpl'))),
+        p.fromUri(packageConfig.resolve(Uri.parse(
+            'package:test/src/runner/browser/static/run_wasm_chrome.js'))),
         root: root);
   }
 
   /// The test runner configuration.
   final Configuration _config;
 
-  /// The underlying server.
-  final shelf.Server _server;
+  /// The cached [CompilerSupport] for each compiler.
+  final _compilerSupport = <Compiler, Future<CompilerSupport>>{};
 
-  /// A randomly-generated secret.
+  /// The `package:test` side wrapper for the Dart2Wasm runtime.
+  final String _jsRuntimeWrapper;
+
+  /// The URL for this server and [compiler] combination.
   ///
-  /// This is used to ensure that other users on the same system can't snoop
-  /// on data being served through this server.
-  final _secret = randomUrlSecret();
-
-  /// The URL for this server.
-  Uri get url => _server.url.resolve('$_secret/');
-
-  /// A [OneOffHandler] for servicing WebSocket connections for
-  /// [BrowserManager]s.
-  ///
-  /// This is one-off because each [BrowserManager] can only connect to a single
-  /// WebSocket,
-  final _webSocketHandler = OneOffHandler();
-
-  /// A [PathHandler] used to serve compiled JS.
-  final _jsHandler = PathHandler();
-
-  /// The [Dart2JsCompilerPool] managing active instances of `dart2js`.
-  final _compilers = Dart2JsCompilerPool();
-
-  /// The temporary directory in which compiled JS is emitted.
-  final String? _compiledDir;
+  /// Each compiler serves its tests under a different randomly-generated
+  /// secret URI to ensure that other users on the same system can't snoop
+  /// on data being served through this server, as well as distinguish tests
+  /// from different compilers from each other.
+  Future<CompilerSupport> compilerSupport(Compiler compiler) =>
+      _compilerSupport.putIfAbsent(compiler, () {
+        if (_config.suiteDefaults.precompiledPath != null) {
+          return PrecompiledSupport.start(
+              compiler: compiler,
+              config: _config,
+              defaultTemplatePath: _defaultTemplatePath,
+              root: _config.suiteDefaults.precompiledPath!,
+              faviconPath: _faviconPath);
+        }
+        return switch (compiler) {
+          Compiler.dart2js => Dart2JsSupport.start(
+              config: _config,
+              defaultTemplatePath: _defaultTemplatePath,
+              root: _root,
+              faviconPath: _faviconPath),
+          Compiler.dart2wasm => Dart2WasmSupport.start(
+              config: _config,
+              defaultTemplatePath: _defaultTemplatePath,
+              jsRuntimeWrapper: _jsRuntimeWrapper,
+              root: _root,
+              faviconPath: _faviconPath),
+          _ => throw StateError('Unexpected compiler $compiler'),
+        };
+      });
 
   /// The root directory served statically by this server.
   final String _root;
-
-  /// The pool of active `pub serve` compilations.
-  ///
-  /// Pub itself ensures that only one compilation runs at a time; we just use
-  /// this pool to make sure that the output is nice and linear.
-  final _pubServePool = Pool(1);
-
-  /// The HTTP client to use when caching JS files in `pub serve`.
-  final HttpClient? _http;
 
   /// Whether [close] has been called.
   bool get _closed => _closeMemo.hasRun;
@@ -109,7 +95,7 @@ class BrowserPlatform extends PlatformPlugin
   /// [BrowserManager]s for those browsers, or `null` if they failed to load.
   ///
   /// This should only be accessed through [_browserManagerFor].
-  final _browserManagers = <Runtime, Future<BrowserManager?>>{};
+  final _browserManagers = <(Runtime, Compiler), Future<BrowserManager?>>{};
 
   /// Settings for invoking each browser.
   ///
@@ -117,71 +103,16 @@ class BrowserPlatform extends PlatformPlugin
   final _browserSettings =
       Map<Runtime, ExecutableSettings>.from(defaultSettings);
 
-  /// A map from test suite paths to Futures that will complete once those
-  /// suites are finished compiling.
-  ///
-  /// This is used to make sure that a given test suite is only compiled once
-  /// per run, rather than once per browser per run.
-  final _compileFutures = <String, Future<void>>{};
-
-  /// Mappers for Dartifying stack traces, indexed by test path.
-  final _mappers = <String, StackTraceMapper>{};
-
   /// The default template for html tests.
   final String _defaultTemplatePath;
 
-  BrowserPlatform._(this._server, Configuration config, String faviconPath,
-      this._defaultTemplatePath,
+  final String _faviconPath;
+
+  BrowserPlatform._(Configuration config, this._faviconPath,
+      this._defaultTemplatePath, this._jsRuntimeWrapper,
       {String? root})
       : _config = config,
-        _root = root ?? p.current,
-        _compiledDir = config.pubServeUrl == null ? createTempDir() : null,
-        _http = config.pubServeUrl == null ? null : HttpClient() {
-    var cascade = shelf.Cascade().add(_webSocketHandler.handler);
-
-    if (_config.pubServeUrl == null) {
-      cascade = cascade
-          .add(packagesDirHandler())
-          .add(_jsHandler.handler)
-          .add(createStaticHandler(
-              config.suiteDefaults.precompiledPath ?? _root,
-              // Precompiled directories often contain symlinks
-              serveFilesOutsidePath:
-                  config.suiteDefaults.precompiledPath != null))
-          .add(_wrapperHandler);
-    }
-
-    var pipeline = shelf.Pipeline()
-        .addMiddleware(PathHandler.nestedIn(_secret))
-        .addHandler(cascade.handler);
-
-    _server.mount(shelf.Cascade()
-        .add(createFileHandler(faviconPath))
-        .add(pipeline)
-        .handler);
-  }
-
-  /// A handler that serves wrapper files used to bootstrap tests.
-  shelf.Response _wrapperHandler(shelf.Request request) {
-    var path = p.fromUri(request.url);
-
-    if (path.endsWith('.html')) {
-      var test = '${p.withoutExtension(path)}.dart';
-      var scriptBase = htmlEscape.convert(p.basename(test));
-      var link = '<link rel="x-dart-test" href="$scriptBase">';
-      var testName = htmlEscape.convert(test);
-      var template = _config.customHtmlTemplatePath ?? _defaultTemplatePath;
-      var contents = File(template).readAsStringSync();
-      var processedContents = contents
-          // Checked during loading phase that there is only one {{testScript}} placeholder.
-          .replaceFirst('{{testScript}}', link)
-          .replaceAll('{{testName}}', testName);
-      return shelf.Response.ok(processedContents,
-          headers: {'Content-Type': 'text/html'});
-    }
-
-    return shelf.Response.notFound('Not found.');
-  }
+        _root = root ?? p.current;
 
   @override
   ExecutableSettings parsePlatformSettings(YamlMap settings) =>
@@ -215,12 +146,7 @@ class BrowserPlatform extends PlatformPlugin
     }
 
     var compiler = platform.compiler;
-    // Technically, ddc is supported through --precompiled. But we don't expect
-    // any specific compiler setting in that case.
-    if (compiler != Compiler.dart2js) {
-      throw StateError('Unsupported compiler for $browser platform '
-          '$compiler');
-    }
+    var support = await compilerSupport(compiler);
 
     var htmlPathFromTestPath = '${p.withoutExtension(path)}.html';
     if (File(htmlPathFromTestPath).existsSync()) {
@@ -248,33 +174,15 @@ class BrowserPlatform extends PlatformPlugin
       _checkHtmlCorrectness(htmlTemplatePath, path);
     }
 
-    Uri suiteUrl;
-    if (_config.pubServeUrl != null) {
-      var suitePrefix = p
-          .toUri(
-              p.withoutExtension(p.relative(path, from: p.join(_root, 'test'))))
-          .path;
+    if (_closed) return null;
+    await support.compileSuite(path, suiteConfig, platform);
 
-      var dartUrl =
-          _config.pubServeUrl!.resolve('$suitePrefix.dart.browser_test.dart');
-
-      await _pubServeSuite(path, dartUrl, browser, suiteConfig);
-      suiteUrl = _config.pubServeUrl!.resolveUri(p.toUri('$suitePrefix.html'));
-    } else {
-      if (suiteConfig.precompiledPath == null) {
-        await _compileSuite(path, suiteConfig);
-      } else {
-        await _addPrecompiledStackTraceMapper(path, suiteConfig);
-      }
-
-      if (_closed) return null;
-      suiteUrl = url.resolveUri(
-          p.toUri('${p.withoutExtension(p.relative(path, from: _root))}.html'));
-    }
+    var suiteUrl = support.serverUrl.resolveUri(
+        p.toUri('${p.withoutExtension(p.relative(path, from: _root))}.html'));
 
     if (_closed) return null;
 
-    var browserManager = await _browserManagerFor(browser);
+    var browserManager = await _browserManagerFor(browser, compiler);
     if (_closed || browserManager == null) return null;
 
     var timeout = const Duration(seconds: 30);
@@ -284,7 +192,8 @@ class BrowserPlatform extends PlatformPlugin
     }
     var suite = await browserManager.load(
         path, suiteUrl, suiteConfig, message, platform.compiler,
-        mapper: _mappers[path], timeout: timeout);
+        mapper: (await compilerSupport(compiler)).stackTraceMapperForPath(path),
+        timeout: timeout);
     if (_closed) return null;
     return suite;
   }
@@ -298,149 +207,29 @@ class BrowserPlatform extends PlatformPlugin
     }
   }
 
-  /// Loads a test suite at [path] from the `pub serve` URL [dartUrl].
-  ///
-  /// This ensures that only one suite is loaded at a time, and that any errors
-  /// are exposed as [LoadException]s.
-  Future<void> _pubServeSuite(String path, Uri dartUrl, Runtime browser,
-      SuiteConfiguration suiteConfig) {
-    return _pubServePool.withResource(() async {
-      var timer = Timer(Duration(seconds: 1), () {
-        print('"pub serve" is compiling $path...');
-      });
-
-      var sourceMapUrl = dartUrl.replace(path: '${dartUrl.path}.js.map');
-
-      try {
-        var request = await _http!.getUrl(sourceMapUrl);
-        var response = await request.close();
-
-        if (response.statusCode != 200) {
-          // Drain response to avoid VM hang.
-          response.drain();
-
-          throw LoadException(
-              path,
-              'Error getting $sourceMapUrl: ${response.statusCode} '
-              '${response.reasonPhrase}\n'
-              'Make sure "pub serve" is serving the test/ directory.');
-        }
-
-        if (suiteConfig.jsTrace) {
-          // Drain response to avoid VM hang.
-          response.drain();
-          return;
-        }
-        _mappers[path] = JSStackTraceMapper(await utf8.decodeStream(response),
-            mapUrl: sourceMapUrl,
-            sdkRoot: p.toUri('packages/\$sdk'),
-            packageMap: (await currentPackageConfig).toPackagesDirPackageMap());
-      } on IOException catch (error) {
-        var message = getErrorMessage(error);
-        if (error is SocketException) {
-          message = '${error.osError?.message} '
-              '(errno ${error.osError?.errorCode})';
-        }
-
-        throw LoadException(
-            path,
-            'Error getting $sourceMapUrl: $message\n'
-            'Make sure "pub serve" is running.');
-      } finally {
-        timer.cancel();
-      }
-    });
-  }
-
-  /// Compile the test suite at [dartPath] to JavaScript.
-  ///
-  /// Once the suite has been compiled, it's added to [_jsHandler] so it can be
-  /// served.
-  Future<void> _compileSuite(String dartPath, SuiteConfiguration suiteConfig) {
-    return _compileFutures.putIfAbsent(dartPath, () async {
-      var dir = Directory(_compiledDir!).createTempSync('test_').path;
-      var jsPath = p.join(dir, '${p.basename(dartPath)}.browser_test.dart.js');
-      var bootstrapContent = '''
-        ${suiteConfig.metadata.languageVersionComment ?? await rootPackageLanguageVersionComment}
-        import "package:test/src/bootstrap/browser.dart";
-
-        import "${p.toUri(p.absolute(dartPath))}" as test;
-
-        void main() {
-          internalBootstrapBrowserTest(() => test.main);
-        }
-      ''';
-
-      await _compilers.compile(bootstrapContent, jsPath, suiteConfig);
-      if (_closed) return;
-
-      var bootstrapUrl = '${p.toUri(p.relative(dartPath, from: _root)).path}'
-          '.browser_test.dart';
-      _jsHandler.add(bootstrapUrl, (request) {
-        return shelf.Response.ok(bootstrapContent,
-            headers: {'Content-Type': 'application/dart'});
-      });
-
-      var jsUrl = '${p.toUri(p.relative(dartPath, from: _root)).path}'
-          '.browser_test.dart.js';
-      _jsHandler.add(jsUrl, (request) {
-        return shelf.Response.ok(File(jsPath).readAsStringSync(),
-            headers: {'Content-Type': 'application/javascript'});
-      });
-
-      var mapUrl = '${p.toUri(p.relative(dartPath, from: _root)).path}'
-          '.browser_test.dart.js.map';
-      _jsHandler.add(mapUrl, (request) {
-        return shelf.Response.ok(File('$jsPath.map').readAsStringSync(),
-            headers: {'Content-Type': 'application/json'});
-      });
-
-      if (suiteConfig.jsTrace) return;
-      var mapPath = '$jsPath.map';
-      _mappers[dartPath] = JSStackTraceMapper(File(mapPath).readAsStringSync(),
-          mapUrl: p.toUri(mapPath),
-          sdkRoot: Uri.parse('org-dartlang-sdk:///sdk'),
-          packageMap: (await currentPackageConfig).toPackageMap());
-    });
-  }
-
-  Future<void> _addPrecompiledStackTraceMapper(
-      String dartPath, SuiteConfiguration suiteConfig) async {
-    if (suiteConfig.jsTrace) return;
-    var mapPath = p.join(
-        suiteConfig.precompiledPath!, '$dartPath.browser_test.dart.js.map');
-    var mapFile = File(mapPath);
-    if (mapFile.existsSync()) {
-      _mappers[dartPath] = JSStackTraceMapper(mapFile.readAsStringSync(),
-          mapUrl: p.toUri(mapPath),
-          sdkRoot: Uri.parse(r'/packages/$sdk'),
-          packageMap: (await currentPackageConfig).toPackageMap());
-    }
-  }
-
   /// Returns the [BrowserManager] for [runtime], which should be a browser.
   ///
   /// If no browser manager is running yet, starts one.
-  Future<BrowserManager?> _browserManagerFor(Runtime browser) {
-    var managerFuture = _browserManagers[browser];
+  Future<BrowserManager?> _browserManagerFor(
+      Runtime browser, Compiler compiler) async {
+    var managerFuture = _browserManagers[(browser, compiler)];
     if (managerFuture != null) return managerFuture;
 
-    var completer = Completer<WebSocketChannel>.sync();
-    var path = _webSocketHandler.create(webSocketHandler(completer.complete));
-    var webSocketUrl = url.replace(scheme: 'ws').resolve(path);
-    var hostUrl = (_config.pubServeUrl ?? url)
+    var support = await compilerSupport(compiler);
+    var (webSocketUrl, socketFuture) = support.webSocket;
+    var hostUrl = support.serverUrl
         .resolve('packages/test/src/runner/browser/static/index.html')
         .replace(queryParameters: {
       'managerUrl': webSocketUrl.toString(),
       'debug': _config.debug.toString()
     });
 
-    var future = BrowserManager.start(browser, hostUrl, completer.future,
-        _browserSettings[browser]!, _config);
+    var future = BrowserManager.start(
+        browser, hostUrl, socketFuture, _browserSettings[browser]!, _config);
 
     // Store null values for browsers that error out so we know not to load them
     // again.
-    _browserManagers[browser] =
+    _browserManagers[(browser, compiler)] =
         future.then<BrowserManager?>((value) => value).onError((_, __) => null);
 
     return future;
@@ -466,19 +255,11 @@ class BrowserPlatform extends PlatformPlugin
   /// Returns a [Future] that completes once the server is closed and its
   /// resources have been fully released.
   @override
-  Future<void> close() async => _closeMemo.runOnce(() async {
-        await Future.wait([
-          for (var browser in _browserManagers.values)
-            browser.then((b) => b?.close()),
-          _server.close(),
-          _compilers.close(),
-        ]);
-
-        if (_config.pubServeUrl == null) {
-          await Directory(_compiledDir!).deleteWithRetry();
-        } else {
-          _http!.close();
-        }
-      });
+  Future<void> close() async => _closeMemo.runOnce(() => Future.wait([
+        for (var browser in _browserManagers.values)
+          browser.then((b) => b?.close()),
+        for (var support in _compilerSupport.values)
+          support.then((s) => s.close()),
+      ]));
   final _closeMemo = AsyncMemoizer<void>();
 }