+++ pkgs/test_core/lib/src/runner/live_suite_controller.dart
@@ -66,12 +66,12 @@ class LiveSuiteController {
   /// The future group that backs [LiveSuite.onComplete].
   ///
   /// This contains all the futures from tests that are run in this suite.
-  final _onCompleteGroup = FutureGroup();
+  final _onCompleteGroup = FutureGroup<void>();
 
   /// The completer that backs [LiveSuite.onClose].
   ///
   /// This is completed when the live suite is closed.
-  final _onCloseCompleter = Completer();
+  final _onCloseCompleter = Completer<void>();
 
   /// The controller for [LiveSuite.onTestStarted].
   final _onTestStartedController =
@@ -150,5 +150,5 @@ class LiveSuiteController {
           _onCloseCompleter.complete();
         }
       });
-  final _closeMemo = AsyncMemoizer();
+  final _closeMemo = AsyncMemoizer<void>();
 }